requirements.txt
pytesseract
pymupdf
pdf2image
numpy
opencv-python
pillow
openai
